+++ src/main/java/org/spring/attraction/service/BoardImage_service.java
@@ -63,28 +63,53 @@ public class BoardImage_service {
     }
 
     //게시글 내 이미지 파일 입력(저장경로 수정, db 내 정보 수정)
-    public void saveImageFile(BoardImage_dto boardImageDto, Board board){
+    public boolean saveImageFile(BoardImage_dto boardImageDto, Board board){
 
         String realUploadPath = UPLOAD_PATH + boardImageDto.getUUIDName();
         String tempOldPath=TEMP_UPLOAD_PATH + boardImageDto.getUUIDName();
+        log.info("temp old path: " + tempOldPath);
+        log.info("real upload path: " + realUploadPath);
+        File tempFile = new File(tempOldPath);
+        File file = new File(realUploadPath);
+
+        if(!tempFile.exists()){
+            log.error("임시 파일이 존재하지 않습니다: {}", tempOldPath);
+            return false;
+        }
 
         try {
-            File tempFile = new File(tempOldPath);
-            File file = new File(realUploadPath);
             file.getParentFile().mkdirs();
+            log.info("파일 이동 진행: {} -> {}",tempFile.toPath(), file.toPath());
             Files.move(tempFile.toPath(), file.toPath(), StandardCopyOption.REPLACE_EXISTING);
         }catch (Exception e) {
             e.printStackTrace();
+            tempFile.delete();
             log.error("파일 이동 실패: {} -> {}", tempOldPath, realUploadPath);
+            return false;
         }
 
         BoardImage boardImage = boardImageRepository.findBoardImageByUUIDAndUUIDName(boardImageDto.getUUID(), boardImageDto.getUUIDName());
 
-        boardImage.setBoard(board);
-        boardImage.setImagePath(realUploadPath);
-        boardImage.setIsTemporary(false);
-        boardImageRepository.save(boardImage);
+        if(boardImage==null){
+            log.error("DB 내 BoardImage 찾기 실패: uuid={}, uuidName={}", boardImageDto.getUUID(), boardImageDto.getUUIDName());
+            return false;
+        }
 
+        try {
+            boardImage.setBoard(board);
+            boardImage.setImagePath(realUploadPath);
+            boardImage.setIsTemporary(false);
+            boardImageRepository.save(boardImage);
+
+        }catch (Exception e) {
+            e.printStackTrace();
+            deleteTempImageFile(boardImage.getUUID(), boardImage.getUUIDName());
+            file.delete();
+            log.error("DB 내 BoardImage 정보 저장실패");
+            return false;
+        }
+
+        return true;
     }
 
     //임시 이미지 파일 읽기
